net: Fix Coverity ID 1508974 Use of 32-bit time_t

+++ source3/utils/net_rpc_sh_acct.c
@@ -355,8 +355,9 @@ static int account_set_minpwage(struct net_context *c,
 	}
 
 	unix_to_nt_time_abs((NTTIME *)&i1->min_password_age, atoi(argv[0]));
-	d_printf(_("Setting minimum password age to %d seconds\n"),
-		 (int)nt_time_to_unix_abs((NTTIME *)&i1->min_password_age));
+	d_printf(_("Setting minimum password age to %" PRIu64 " seconds\n"),
+		 (uint64_t)nt_time_to_unix_abs(
+			 (NTTIME *)&i1->min_password_age));
 
 	return 1;
 }
